Returned from Sets back to Lists for storing the child relationships and returned to fetch type LAZY for one-to-many relationships. Turns out this combination works perfectly and due to LAZY fetch is also faster.
Due to these changes deleting children now works perfectly with deleteById, so replaced deleteDownProp with it.

Added annotation such that the children list is ordered and the order is persisted in the database. This is essential to drag&drop functionality.

Replaced building BAD_REQUESTs and requests NOT_FOUND with throwing an exception. This exception is automatically handled by Spring and allows to include a message with the error.

Changed CardList field for holding cards from "cardList" to "cards" because I think it's less confusing.

Adjusted the controller tests to the changes.

Signed-off-by: blytkowski <b.lytkowski@student.tudelft.nl>